update forking interface
