+++ crates/cheatnet/src/forking.rs
@@ -1,2 +1,1 @@
 pub mod state;
-pub mod worker;
